pyproject.toml
[tool.poetry]
name = "usc-discord-bot"
version = "0.1.0"
description = ""
authors = ["Zachary Bryant <Zachary@admiralfeb.net>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.1"
discord-py-slash-command = "^1.1.2"
python-dotenv = "^0.17.0"
pymongo = {extras = ["srv"], version = "^3.11.3"}
bugsnag = "^4.0.2"
requests = "^2.25.1"
"quickchart.io" = "^0.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.7.4"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
